chore: new folder structure
