+++ features/step_definitions/api/auth.steps.js
@@ -1,0 +1,9 @@
+const { When, Then, Given } = require("@cucumber/cucumber");
+const { expect } = require("@playwright/test");
+
+Given(
+  "I authenticated as username {string} and password {string}",
+  async function (username, password) {
+    this.currentAuth = this.getAuthHeader(username, password);
+  }
+);
