+++ features/step_definitions/api/book.step.js
@@ -1,47 +1,5 @@
-const { Given, When, Then } = require("@cucumber/cucumber");
+const { When, Then } = require("@cucumber/cucumber");
 const { expect } = require("@playwright/test");
-const serverUtils = require("../../support/server-utils");
-
-Given("the book database is empty", async function () {
-  const cleaned = await serverUtils.clearDatabase();
-  expect(cleaned).toBe(true);
-});
-
-Given(
-  "I authenticated as username {string} and password {string}",
-  async function (username, password) {
-    this.currentAuth = serverUtils.getAuthHeader(username, password);
-  }
-);
-
-When(
-  "I send a {string} request to {string}",
-  async function (method, endpoint) {
-    // Replace stored ID if present in endpoint
-    const finalEndpoint = this.storedBookId
-      ? endpoint.replace("{stored-id}", this.storedBookId)
-      : endpoint;
-
-    const headers = {
-      ...this.currentAuth,
-    };
-
-    switch (method.toUpperCase()) {
-      case "GET":
-        this.response = await this.apiContext.get(finalEndpoint, {
-          headers,
-        });
-        break;
-      case "DELETE":
-        this.response = await this.apiContext.delete(finalEndpoint, {
-          headers,
-        });
-        break;
-      default:
-        throw new Error(`Unsupported HTTP method: ${method}`);
-    }
-  }
-);
 
 When(
   "I have created a book with title {string} and author {string}",
@@ -63,6 +21,12 @@ When(
     }
   }
 );
+
+Then("the response should contain an empty list", async function () {
+  const books = await this.response.json();
+  expect(Array.isArray(books)).toBe(true);
+  expect(books.length).toBe(0);
+});
 
 When(
   "I send a {string} request to {string} with updated details:",
@@ -95,12 +59,6 @@ When(
   }
 );
 
-Then("the response should contain an empty list", async function () {
-  const books = await this.response.json();
-  expect(Array.isArray(books)).toBe(true);
-  expect(books.length).toBe(0);
-});
-
 Then(
   "the response should contain the created book details title {string} and author {string}",
   async function (expectedTitle, expectedAuthor) {
@@ -112,6 +70,9 @@ Then(
 
 Then("I store the created book ID", async function () {
   const responseData = await this.response.json();
-  this.storedBookId = responseData.id;
-  expect(this.storedBookId).toBeDefined();
+
+  if (responseData.id) {
+    this.storedBookId = responseData.id;
+    expect(this.storedBookId).toBeDefined();
+  }
 });
