+++ features/step_definitions/api/common.steps.js
@@ -1,0 +1,80 @@
+const { When, Then, Given } = require("@cucumber/cucumber");
+const { expect } = require("@playwright/test");
+const serverUtils = require("../../support/server-utils");
+
+Then("the response status code should be {int}", async function (statusCode) {
+  expect(this.response.status()).toBe(statusCode);
+});
+
+Given("the book database is empty", async function () {
+  const cleaned = await serverUtils.clearDatabase();
+  expect(cleaned).toBe(true);
+});
+
+When(
+  "I send a {string} request to {string}",
+  async function (method, endpoint) {
+    // Replace stored ID if present in endpoint
+    const finalEndpoint = this.storedBookId
+      ? endpoint.replace("{stored-id}", this.storedBookId)
+      : endpoint;
+
+    const headers = {
+      ...this.currentAuth,
+    };
+
+    switch (method.toUpperCase()) {
+      case "GET":
+        this.response = await this.apiContext.get(finalEndpoint, {
+          headers,
+        });
+        break;
+      case "DELETE":
+        this.response = await this.apiContext.delete(finalEndpoint, {
+          headers,
+        });
+        break;
+      default:
+        throw new Error(`Unsupported HTTP method: ${method}`);
+    }
+  }
+);
+
+When(
+  "I send a {string} request to {string} without authentication",
+  async function (method, endpoint) {
+    // EXPLICTLY SETTING HEADERS TO EMPTY OBJECT
+    const headers = {};
+
+    switch (method.toUpperCase()) {
+      case "GET":
+        this.response = await this.apiContext.get(endpoint, {
+          headers,
+        });
+        break;
+      case "POST":
+        this.response = await this.apiContext.post(endpoint, {
+          headers,
+          data: {
+            title: "Test Book",
+            author: "Test Author",
+          },
+        });
+        break;
+      case "PUT":
+        this.response = await this.apiContext.put(endpoint, {
+          headers,
+          data: {
+            title: "Updated Book",
+            author: "Updated Author",
+          },
+        });
+        break;
+      case "DELETE":
+        this.response = await this.apiContext.delete(endpoint, {
+          headers,
+        });
+        break;
+    }
+  }
+);
